refactored and finished queues
